requirements.txt
pyspacemouse
pihat
flask==3.0.3
paho-mqtt==2.1.0
pyyaml==6.0.1
pydantic==2.7.1